Refactor for Running Each Test Cases
