+++ .env
@@ -10,8 +10,8 @@ TRACE_FORMAT="{{4}} - {{5}}   .   {{6}}"
 COMMENT_FORMAT="{{1}} {{2}}"
 # ACTION_FORMAT="{{1}} - {{2}}"
 
-FILE="sample-tests.xlsx"
-SHEET=17
+FILE=".\sample-tests\sample-tests.xlsx"
+SHEET=1,2
 
 # ------------------------------------------------------------------------------
 # You can have any other cutom key=value pairs here...
