+++ .gitignore
@@ -1,2 +1,10 @@
 node_modules/
 _
+
+# Playwright
+/test-results/
+/playwright-report/
+/blob-report/
+/playwright/.cache/
+/output.log
+/tests-generatedfiles/
