+++ package.json
@@ -1,19 +1,18 @@
 {
-  "name": "e2erunner2",
-  "version": "2.0.0",
-  "description": "",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
+  "name": "ibot",
+  "version": "2.16.0",
+  "description": "Test Automation Framework with Playwright",
+  "scripts": {},
   "keywords": [],
   "author": "",
   "license": "ISC",
   "dependencies": {
-    "@playwright/test": "^1.22.2",
-    "dotenv": "^16.0.1",
-    "exceljs": "^4.3.0",
-    "handlebars": "^4.7.7",
-    "moment": "^2.29.4"
+    "@playwright/test": "^1.50.1",
+    "dotenv": "^16.4.7",
+    "e2erunner2": "file:",
+    "exceljs": "^4.4.0",
+    "handlebars": "^4.7.8",
+    "ibot": "file:",
+    "moment": "^2.30.1"
   }
 }
