+++ playwright.config.ts
@@ -2,7 +2,7 @@ import { PlaywrightTestConfig } from '@playwright/test';
 
 const config: PlaywrightTestConfig = {
   reportSlowTests: null,
-  timeout: 1000 * 60 * 10, //10 minutes
+  //timeout: 1000 * 60 * 10, //10 minutes
   // use: {
   //   video: 'on',
   //   screenshot: 'on',
