+++ tests/main.spec.ts
@@ -1,12 +1,12 @@
 import { test } from '@playwright/test'
 import { Workbook } from 'exceljs'
-import { runSheet } from './actions'
+import { runSheet } from '../src/actions'
 import {
   ACTION, ACTION_FORMAT, COMMENT_FORMAT,
   DATA, FILE, humanNowDateTime, LOCATOR, PRINT_FORMAT,
   SHEET, TRACE, TRACE_FORMAT,
-} from './consts'
-import { logAll, logSheetClose, parseInts, SHEET_TIMER, TOTAL_SUMMARY, TOTAL_TIMER } from './lib'
+} from '../src/consts'
+import { logAll, logSheetClose, parseInts, SHEET_TIMER, TOTAL_SUMMARY, TOTAL_TIMER } from './src/lib'
 
 test('check all', async ({ page, context }) => {
   // console.log(context.br)
@@ -57,4 +57,3 @@ test('check all', async ({ page, context }) => {
   logAll()
 
 })
-
